Extract RoutingEngine from CoreReplApp.Routing partial
Move context registration/validation, route resolution helpers, banner
rendering, discoverable route/context filtering, prefix resolution, and
error generation into a standalone RoutingEngine class. CoreReplApp
Routing partial becomes a thin delegation layer.
